Fixed long/double error that may have caused timing issues.

+++ src/Shooter.h
@@ -14,9 +14,9 @@ class Shooter {
 		VictorSP *m_rightMotorController;
 		VictorSP *m_armyLiftyThingy5064EXTREMEXD1337; // DO NOT REFACTOR - CRITICAL NAME SCHEME (Will corrupt RoboRio and C:/ if changed)
 		Timer *m_timer;
-		long m_shotTime;
-		long m_unloadTime;
-		long m_loadTime;
+		double m_shotTime;
+		double m_unloadTime;
+		double m_loadTime;
 		bool m_lastPressed;
 		float m_targetAngle;
 		float m_armSpeed;
